even vs odd ratio demo
this script is meant to show a demo of the almost perfect 2:1 even:odd results in the collatz conjecture, which explains the dramatic decrease in sum in each iteration of the equation. because odds always result in an even result in this conjecture, whereas evens have a chance to result in evens or odds, there is an obvious imbalance in the equation which will statistically result in the sum always dropping to 1. this is not a proof, more of a demo concept of the idea that the balance is pretty obviously leaning toward even results (which obviously will result in that number being divided by 2, thus reducing the number) which always brings the result lower and lower until it hits 1. 

sorry if this is a bad explanation, im new to all this math stuff, but writing this and watching the ratio bounce ever so slightly below (and sometimes ever so slightly above) 2:1 is really satisfying to me so I wanted to share